+++ saleor/graphql/core/mutations.py
@@ -27,7 +27,13 @@ from ...core.permissions import (
 )
 from ..utils import get_nodes, resolve_global_ids_to_primary_keys
 from .descriptions import DEPRECATED_IN_3X_FIELD
-from .types import File, NonNullList, Upload, UploadError
+from .types import (
+    TYPES_WITH_DOUBLE_ID_AVAILABLE,
+    File,
+    NonNullList,
+    Upload,
+    UploadError,
+)
 from .utils import from_global_id_or_error, snake_to_camel_case
 from .utils.error_codes import get_error_code_from_error
 
@@ -177,25 +183,23 @@ class BaseMutation(graphene.Mutation):
         Whether by using the provided query set object or by calling type's get_node().
         """
         if qs is not None:
-            if str(graphene_type) == "Order":
-                return cls._get_order_node_by_pk(pk, qs)
-            return qs.filter(pk=pk).first()
+            lookup = Q(pk=pk)
+            if pk is not None and str(graphene_type) in TYPES_WITH_DOUBLE_ID_AVAILABLE:
+                # This is temporary solution that allows fetching objects with use of
+                # new and old id.
+                try:
+                    UUID(str(pk))
+                except ValueError:
+                    lookup = (
+                        Q(number=pk) & Q(use_old_id=True)
+                        if str(graphene_type) == "Order"
+                        else Q(old_id=pk) & Q(old_id__isnull=False)
+                    )
+            return qs.filter(lookup).first()
         get_node = getattr(graphene_type, "get_node", None)
         if get_node:
             return get_node(info, pk)
         return None
-
-    @classmethod
-    def _get_order_node_by_pk(cls, pk: Union[int, str], qs):
-        # This is temporary method that allows fetching orders with use of
-        # new and old id.
-        lookup = Q(pk=pk)
-        if pk is not None:
-            try:
-                UUID(str(pk))
-            except ValueError:
-                lookup = Q(number=pk) & Q(use_old_id=True)
-        return qs.filter(lookup).first()
 
     @classmethod
     def get_global_id_or_error(