+++ saleor/graphql/core/types/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 from .common import (
+    TYPES_WITH_DOUBLE_ID_AVAILABLE,
     AccountError,
     AppError,
     AttributeError,