+++ saleor/graphql/webhook/subscription_types.py
@@ -38,6 +38,34 @@ TRANSLATIONS_TYPES_MAP = {
 }
 
 
+class AppBase(AbstractType):
+    app = graphene.Field(
+        "saleor.graphql.app.types.App",
+        description="Look up a app." + ADDED_IN_34 + PREVIEW_FEATURE,
+    )
+
+    @staticmethod
+    def resolve_app(root, _info):
+        _, app = root
+        return app
+
+
+class AppCreated(ObjectType, AppBase):
+    ...
+
+
+class AppUpdated(ObjectType, AppBase):
+    ...
+
+
+class AppDeleted(ObjectType, AppBase):
+    ...
+
+
+class AppStatusChanged(ObjectType, AppBase):
+    ...
+
+
 class CategoryBase(AbstractType):
     category = graphene.Field(
         "saleor.graphql.product.types.Category",
@@ -163,6 +191,60 @@ class GiftCardDeleted(ObjectType, GiftCardBase):
 
 
 class GiftCardStatusChanged(ObjectType, GiftCardBase):
+    ...
+
+
+class MenuBase(AbstractType):
+    menu = graphene.Field(
+        "saleor.graphql.menu.types.Menu",
+        channel=graphene.String(
+            description="Slug of a channel for which the data should be returned."
+        ),
+        description="Look up a menu." + ADDED_IN_34 + PREVIEW_FEATURE,
+    )
+
+    @staticmethod
+    def resolve_menu(root, info, channel=None):
+        _, menu = root
+        return ChannelContext(node=menu, channel_slug=channel)
+
+
+class MenuCreated(ObjectType, MenuBase):
+    ...
+
+
+class MenuUpdated(ObjectType, MenuBase):
+    ...
+
+
+class MenuDeleted(ObjectType, MenuBase):
+    ...
+
+
+class MenuItemBase(AbstractType):
+    menu_item = graphene.Field(
+        "saleor.graphql.menu.types.MenuItem",
+        channel=graphene.String(
+            description="Slug of a channel for which the data should be returned."
+        ),
+        description="Look up a menu item." + ADDED_IN_34 + PREVIEW_FEATURE,
+    )
+
+    @staticmethod
+    def resolve_menu_item(root, info, channel=None):
+        _, menu_item = root
+        return ChannelContext(node=menu_item, channel_slug=channel)
+
+
+class MenuItemCreated(ObjectType, MenuItemBase):
+    ...
+
+
+class MenuItemUpdated(ObjectType, MenuItemBase):
+    ...
+
+
+class MenuItemDeleted(ObjectType, MenuItemBase):
     ...
 
 
@@ -597,6 +679,10 @@ class VoucherDeleted(ObjectType, VoucherBase):
 class Event(Union):
     class Meta:
         types = (
+            AppCreated,
+            AppUpdated,
+            AppDeleted,
+            AppStatusChanged,
             CategoryCreated,
             CategoryUpdated,
             CategoryDeleted,
@@ -608,6 +694,12 @@ class Event(Union):
             GiftCardUpdated,
             GiftCardDeleted,
             GiftCardStatusChanged,
+            MenuCreated,
+            MenuUpdated,
+            MenuDeleted,
+            MenuItemCreated,
+            MenuItemUpdated,
+            MenuItemDeleted,
             OrderCreated,
             OrderUpdated,
             OrderConfirmed,
@@ -660,6 +752,10 @@ class Event(Union):
     @classmethod
     def get_type(cls, object_type: str):
         types = {
+            WebhookEventAsyncType.APP_CREATED: AppCreated,
+            WebhookEventAsyncType.APP_UPDATED: AppUpdated,
+            WebhookEventAsyncType.APP_DELETED: AppDeleted,
+            WebhookEventAsyncType.APP_STATUS_CHANGED: AppStatusChanged,
             WebhookEventAsyncType.CATEGORY_CREATED: CategoryCreated,
             WebhookEventAsyncType.CATEGORY_UPDATED: CategoryUpdated,
             WebhookEventAsyncType.CATEGORY_DELETED: CategoryDeleted,
@@ -671,6 +767,12 @@ class Event(Union):
             WebhookEventAsyncType.GIFT_CARD_UPDATED: GiftCardUpdated,
             WebhookEventAsyncType.GIFT_CARD_DELETED: GiftCardDeleted,
             WebhookEventAsyncType.GIFT_CARD_STATUS_CHANGED: GiftCardStatusChanged,
+            WebhookEventAsyncType.MENU_CREATED: MenuCreated,
+            WebhookEventAsyncType.MENU_UPDATED: MenuUpdated,
+            WebhookEventAsyncType.MENU_DELETED: MenuDeleted,
+            WebhookEventAsyncType.MENU_ITEM_CREATED: MenuItemCreated,
+            WebhookEventAsyncType.MENU_ITEM_UPDATED: MenuItemUpdated,
+            WebhookEventAsyncType.MENU_ITEM_DELETED: MenuItemDeleted,
             WebhookEventAsyncType.ORDER_CREATED: OrderCreated,
             WebhookEventAsyncType.ORDER_UPDATED: OrderUpdated,
             WebhookEventAsyncType.ORDER_CONFIRMED: OrderConfirmed,