+++ saleor/plugins/webhook/plugin.py
@@ -57,6 +57,7 @@ if TYPE_CHECKING:
     from ...giftcard.models import GiftCard
     from ...graphql.discount.mutations import NodeCatalogueInfo
     from ...invoice.models import Invoice
+    from ...menu.models import Menu, MenuItem
     from ...order.models import Fulfillment, Order
     from ...page.models import Page
     from ...payment.interface import (
@@ -91,6 +92,35 @@ class WebhookPlugin(BasePlugin):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.active = True
+
+    def _trigger_app_event(self, event_type, app):
+        if webhooks := get_webhooks_for_event(event_type):
+            payload = {
+                "id": graphene.Node.to_global_id("App", app.id),
+                "is_active": app.is_active,
+                "name": app.name,
+            }
+            trigger_webhooks_async(payload, event_type, webhooks, app, self.requestor)
+
+    def app_created(self, app: "App", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_app_event(WebhookEventAsyncType.APP_CREATED, app)
+
+    def app_updated(self, app: "App", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_app_event(WebhookEventAsyncType.APP_UPDATED, app)
+
+    def app_deleted(self, app: "App", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_app_event(WebhookEventAsyncType.APP_DELETED, app)
+
+    def app_status_changed(self, app: "App", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_app_event(WebhookEventAsyncType.APP_STATUS_CHANGED, app)
 
     def category_created(self, category: "Category", previous_value: None) -> None:
         if not self.active:
@@ -214,6 +244,61 @@ class WebhookPlugin(BasePlugin):
             trigger_webhooks_async(
                 order_data, event_type, webhooks, order, self.requestor
             )
+
+    def _trigger_menu_event(self, event_type, menu):
+        if webhooks := get_webhooks_for_event(event_type):
+            payload = {
+                "id": graphene.Node.to_global_id("Menu", menu.id),
+                "slug": menu.slug,
+            }
+            trigger_webhooks_async(payload, event_type, webhooks, menu, self.requestor)
+
+    def menu_created(self, menu: "Menu", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_menu_event(WebhookEventAsyncType.MENU_CREATED, menu)
+
+    def menu_updated(self, menu: "Menu", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_menu_event(WebhookEventAsyncType.MENU_UPDATED, menu)
+
+    def menu_deleted(self, menu: "Menu", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_menu_event(WebhookEventAsyncType.MENU_DELETED, menu)
+
+    def __trigger_menu_item_event(self, event_type, menu_item):
+        if webhooks := get_webhooks_for_event(event_type):
+            payload = {
+                "id": graphene.Node.to_global_id("MenuItem", menu_item.id),
+                "name": menu_item.name,
+                "menu": {"id": graphene.Node.to_global_id("Menu", menu_item.menu_id)},
+            }
+            trigger_webhooks_async(
+                payload, event_type, webhooks, menu_item, self.requestor
+            )
+
+    def menu_item_created(self, menu_item: "MenuItem", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self.__trigger_menu_item_event(
+            WebhookEventAsyncType.MENU_ITEM_CREATED, menu_item
+        )
+
+    def menu_item_updated(self, menu_item: "MenuItem", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self.__trigger_menu_item_event(
+            WebhookEventAsyncType.MENU_ITEM_UPDATED, menu_item
+        )
+
+    def menu_item_deleted(self, menu_item: "MenuItem", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self.__trigger_menu_item_event(
+            WebhookEventAsyncType.MENU_ITEM_DELETED, menu_item
+        )
 
     def order_confirmed(self, order: "Order", previous_value: Any) -> Any:
         if not self.active: